refactor(auth): move SessionRepository interface to application/usecases
C3 из 6-step ADR-6 sequence. Mechanical mirror C2: file rename
domain/repositories → application/usecases, 2 callers (session_repository
PG impl + integration test suite) обновили alias `repositories` →
`usecases`.

Build / test / lint зелёные.
